+++ bin/domain/usecases/ClearNextActionHourUseCase.js
@@ -6,9 +6,10 @@ class ClearNextActionHourUseCase {
         this.issueRepository = issueRepository;
         this.run = async (input) => {
             const nextActionHour = input.project.nextActionHour;
-            if (!nextActionHour || input.cacheUsed) {
+            if (!nextActionHour) {
                 return;
             }
+            const nextActionDate = input.project.nextActionDate;
             const targetDates = input.targetDates
                 .filter((targetDate) => targetDate.getMinutes() === 45)
                 .reverse();
@@ -16,7 +17,7 @@ class ClearNextActionHourUseCase {
                 return;
             }
             const targetDate = new Date(targetDates[targetDates.length - 1].getTime() + 5 * 60 * 1000);
-            const targetHour = targetDate.getHours();
+            const targetHour = targetDate.getHours() + 1;
             const isTargetIssue = (issue) => {
                 return (issue.nextActionHour !== null &&
                     issue.nextActionHour <= targetHour &&
@@ -29,6 +30,11 @@ class ClearNextActionHourUseCase {
                     continue;
                 }
                 await this.issueRepository.clearProjectField(input.project, nextActionHour.fieldId, issue);
+                await new Promise((resolve) => setTimeout(resolve, 5000));
+                if (!nextActionDate) {
+                    continue;
+                }
+                await this.issueRepository.clearProjectField(input.project, nextActionDate.fieldId, issue);
                 await new Promise((resolve) => setTimeout(resolve, 5000));
             }
         };