+++ types/domain/usecases/HandleScheduledEventUseCase.d.ts
@@ -13,91 +13,77 @@ import { AnalyzeProblemByIssueUseCase } from './AnalyzeProblemByIssueUseCase';
 import { AnalyzeStoriesUseCase } from './AnalyzeStoriesUseCase';
 import { ClearDependedIssueURLUseCase } from './ClearDependedIssueURLUseCase';
 import { CreateEstimationIssueUseCase } from './CreateEstimationIssueUseCase';
+import { ConvertCheckboxToIssueInStoryIssueUseCase } from './ConvertCheckboxToIssueInStoryIssueUseCase';
 export declare class ProjectNotFoundError extends Error {
-  constructor(message: string);
+    constructor(message: string);
 }
 export type StoryObject = {
-  story: StoryOption;
-  storyIssue: Issue | null;
-  issues: (Issue & {
-    totalWorkingTime: number;
-    totalWorkingTimeByAssignee: Map<string, number>;
-  })[];
+    story: StoryOption;
+    storyIssue: Issue | null;
+    issues: (Issue & {
+        totalWorkingTime: number;
+        totalWorkingTimeByAssignee: Map<string, number>;
+    })[];
 };
 export type StoryObjectMap = Map<string, StoryObject>;
 export declare class HandleScheduledEventUseCase {
-  readonly generateWorkingTimeReportUseCase: GenerateWorkingTimeReportUseCase;
-  readonly actionAnnouncementUseCase: ActionAnnouncementUseCase;
-  readonly setWorkflowManagementIssueToStoryUseCase: SetWorkflowManagementIssueToStoryUseCase;
-  readonly clearNextActionHourUseCase: ClearNextActionHourUseCase;
-  readonly analyzeProblemByIssueUseCase: AnalyzeProblemByIssueUseCase;
-  readonly analyzeStoriesUseCase: AnalyzeStoriesUseCase;
-  readonly clearDependedIssueURLUseCase: ClearDependedIssueURLUseCase;
-  readonly createEstimationIssueUseCase: CreateEstimationIssueUseCase;
-  readonly dateRepository: DateRepository;
-  readonly spreadsheetRepository: SpreadsheetRepository;
-  readonly projectRepository: ProjectRepository;
-  readonly issueRepository: IssueRepository;
-  constructor(
-    generateWorkingTimeReportUseCase: GenerateWorkingTimeReportUseCase,
-    actionAnnouncementUseCase: ActionAnnouncementUseCase,
-    setWorkflowManagementIssueToStoryUseCase: SetWorkflowManagementIssueToStoryUseCase,
-    clearNextActionHourUseCase: ClearNextActionHourUseCase,
-    analyzeProblemByIssueUseCase: AnalyzeProblemByIssueUseCase,
-    analyzeStoriesUseCase: AnalyzeStoriesUseCase,
-    clearDependedIssueURLUseCase: ClearDependedIssueURLUseCase,
-    createEstimationIssueUseCase: CreateEstimationIssueUseCase,
-    dateRepository: DateRepository,
-    spreadsheetRepository: SpreadsheetRepository,
-    projectRepository: ProjectRepository,
-    issueRepository: IssueRepository,
-  );
-  run: (input: {
-    projectName: string;
-    org: string;
-    projectUrl: string;
-    manager: Member['name'];
-    workingReport: {
-      repo: string;
-      members: Member['name'][];
-      warningThresholdHour?: number;
-      spreadsheetUrl: string;
-      reportIssueTemplate?: string;
-      reportIssueLabels: Label[];
-    };
-    urlOfStoryView: string;
-    disabledStatus: string;
-  }) => Promise<{
-    project: Project;
-    issues: Issue[];
-    cacheUsed: boolean;
-    targetDateTimes: Date[];
-    storyIssues: StoryObjectMap;
-  }>;
-  runForTargetDateTime: (input: {
-    org: string;
-    manager: Member['name'];
-    workingReport: {
-      repo: string;
-      members: Member['name'][];
-      warningThresholdHour?: number;
-      spreadsheetUrl: string;
-      reportIssueTemplate?: string;
-      reportIssueLabels: Label[];
-    };
-    projectId: Project['id'];
-    issues: Issue[];
-    targetDateTime: Date;
-  }) => Promise<void>;
-  static createTargetDateTimes: (from: Date, to: Date) => Date[];
-  findTargetDateAndUpdateLastExecutionDateTime: (
-    spreadsheetUrl: string,
-    now: Date,
-  ) => Promise<Date[]>;
-  storyIssues: (input: {
-    project: Project;
-    issues: Issue[];
-  }) => Promise<StoryObjectMap>;
-  calculateTotalWorkingMinutesByAssignee: (issue: Issue) => Map<string, number>;
+    readonly generateWorkingTimeReportUseCase: GenerateWorkingTimeReportUseCase;
+    readonly actionAnnouncementUseCase: ActionAnnouncementUseCase;
+    readonly setWorkflowManagementIssueToStoryUseCase: SetWorkflowManagementIssueToStoryUseCase;
+    readonly clearNextActionHourUseCase: ClearNextActionHourUseCase;
+    readonly analyzeProblemByIssueUseCase: AnalyzeProblemByIssueUseCase;
+    readonly analyzeStoriesUseCase: AnalyzeStoriesUseCase;
+    readonly clearDependedIssueURLUseCase: ClearDependedIssueURLUseCase;
+    readonly createEstimationIssueUseCase: CreateEstimationIssueUseCase;
+    readonly convertCheckboxToIssueInStoryIssueUseCase: ConvertCheckboxToIssueInStoryIssueUseCase;
+    readonly dateRepository: DateRepository;
+    readonly spreadsheetRepository: SpreadsheetRepository;
+    readonly projectRepository: ProjectRepository;
+    readonly issueRepository: IssueRepository;
+    constructor(generateWorkingTimeReportUseCase: GenerateWorkingTimeReportUseCase, actionAnnouncementUseCase: ActionAnnouncementUseCase, setWorkflowManagementIssueToStoryUseCase: SetWorkflowManagementIssueToStoryUseCase, clearNextActionHourUseCase: ClearNextActionHourUseCase, analyzeProblemByIssueUseCase: AnalyzeProblemByIssueUseCase, analyzeStoriesUseCase: AnalyzeStoriesUseCase, clearDependedIssueURLUseCase: ClearDependedIssueURLUseCase, createEstimationIssueUseCase: CreateEstimationIssueUseCase, convertCheckboxToIssueInStoryIssueUseCase: ConvertCheckboxToIssueInStoryIssueUseCase, dateRepository: DateRepository, spreadsheetRepository: SpreadsheetRepository, projectRepository: ProjectRepository, issueRepository: IssueRepository);
+    run: (input: {
+        projectName: string;
+        org: string;
+        projectUrl: string;
+        manager: Member["name"];
+        workingReport: {
+            repo: string;
+            members: Member["name"][];
+            warningThresholdHour?: number;
+            spreadsheetUrl: string;
+            reportIssueTemplate?: string;
+            reportIssueLabels: Label[];
+        };
+        urlOfStoryView: string;
+        disabledStatus: string;
+    }) => Promise<{
+        project: Project;
+        issues: Issue[];
+        cacheUsed: boolean;
+        targetDateTimes: Date[];
+        storyIssues: StoryObjectMap;
+    }>;
+    runForTargetDateTime: (input: {
+        org: string;
+        manager: Member["name"];
+        workingReport: {
+            repo: string;
+            members: Member["name"][];
+            warningThresholdHour?: number;
+            spreadsheetUrl: string;
+            reportIssueTemplate?: string;
+            reportIssueLabels: Label[];
+        };
+        projectId: Project["id"];
+        issues: Issue[];
+        targetDateTime: Date;
+    }) => Promise<void>;
+    static createTargetDateTimes: (from: Date, to: Date) => Date[];
+    findTargetDateAndUpdateLastExecutionDateTime: (spreadsheetUrl: string, now: Date) => Promise<Date[]>;
+    storyIssues: (input: {
+        project: Project;
+        issues: Issue[];
+    }) => Promise<StoryObjectMap>;
+    calculateTotalWorkingMinutesByAssignee: (issue: Issue) => Map<string, number>;
 }
 //# sourceMappingURL=HandleScheduledEventUseCase.d.ts.map